+++ tests/Config/ConfigFile/ConfigFileTest.php
@@ -67,6 +67,6 @@ class ConfigFileTest extends PHPUnit_Framework_TestCase
         $baseConfig->set(Config::PULI_DIR, 'my-puli-dir');
 
         $this->assertSame('my-puli-dir', $config->get(Config::PULI_DIR));
-        $this->assertNull($config->get(Config::PULI_DIR, false));
+        $this->assertNull($config->get(Config::PULI_DIR, null, false));
     }
 }