+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Changelog
  * renamed `installPath` to `install-path` in puli.json
  * added `bindings` and `binding-types` keys to puli.json
  * removed `tags` and `tag-definitions` keys from puli.json
+ * added default returned value to `Config::get()` and `Config::getRaw()` 
 
 * 1.0.0-alpha1 (2014-12-03)